# auth-api/openapi.json
{
    "openapi": "3.0.0",
    "info": {
      "title": "Auth API",
      "version": "1.0.0",
      "description": "A secured API with sign up, log in, log out, and protected routes, backed by Neon Auth."
    },
    "servers": [
      { "url": "http://localhost:3000" }
    ],
    "components": {
      "securitySchemes": {
        "bearerAuth": {
          "type": "http",
          "scheme": "bearer",
          "bearerFormat": "JWT"
        }
      }
    },
    "paths": {
      "/public/info": {
        "get": {
          "summary": "Public info",
          "responses": {
            "200": { "description": "Public message" }
          }
        }
      },
      "/auth/signup": {
        "post": {
          "summary": "Create a new user account",
          "requestBody": {
            "required": true,
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "email": { "type": "string" },
                    "password": { "type": "string" }
                  },
                  "required": ["email", "password"]
                }
              }
            }
          },
          "responses": {
            "201": { "description": "User created" },
            "400": { "description": "Missing email or password" }
          }
        }
      },
      "/auth/login": {
        "post": {
          "summary": "Authenticate and return a JWT",
          "requestBody": {
            "required": true,
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "email": { "type": "string" },
                    "password": { "type": "string" }
                  },
                  "required": ["email", "password"]
                }
              }
            }
          },
          "responses": {
            "200": { "description": "Login successful, returns access_token" },
            "400": { "description": "Missing email or password" },
            "401": { "description": "Invalid login credentials" }
          }
        }
      },
      "/auth/logout": {
        "post": {
          "summary": "End the user's session",
          "security": [{ "bearerAuth": [] }],
          "responses": {
            "204": { "description": "Logged out" },
            "401": { "description": "Missing or invalid token" }
          }
        }
      },
      "/protected/profile": {
        "get": {
          "summary": "Read private profile data",
          "security": [{ "bearerAuth": [] }],
          "responses": {
            "200": { "description": "User profile" },
            "401": { "description": "Missing, invalid, or expired token" }
          }
        }
      },
      "/protected/dashboard": {
        "get": {
          "summary": "Read protected dashboard data",
          "security": [{ "bearerAuth": [] }],
          "responses": {
            "200": { "description": "Dashboard welcome message" },
            "401": { "description": "Missing, invalid, or expired token" }
          }
        }
      }
    }
  }